2017-11-30-new-17-profiles: Display based on profile, not gcc version
Display the news item on 17.0 profile upgrade to all users using 13.0
profiles, rather than the people having new enough gcc installed. Given
that all users see the new profile, it is not a good idea to show
the news item to subset of them and leave the rest without any
instructions. Furthermore, displaying the item to users using 17.0
already is usually pointless.