net-proxy/squid: fix cross-compiling
Current versions of squid have a run test for gnu atomics w/no
cross-compile fallback causing builds to fail.  Upstream has
rewritten this completely, so until we update to a newer version,
force the cache variable when cross-compiling.  Forcing it to no
might not be optimal in all cases, but it's better than nothing.

We have been turning on large-file-system support for a while now,
but the configure test for that probes the build system for flags
to compile with (via `getconf ...`).  These flags don't work for
the cross-compile target, and in reality are rarely necessary.  By
selecting the default, we always go the standard POSIX route with
the -D_FILE_OFFSET_BITS=64 flag.  If a target doesn't respect that
flag, then it might fail to build, but we can wait for anyone to
complain.