app-portage/unsymlink-lib: Bump to v11, with lib32 removal fix
Bump to v11 that improves handling of lib32 removal failures. The errors
are now fatal and request resuming explicitly, rather than outputting
a warning that is easily missed. This also lets us allow lib32
to be a mount point again.

This version also tries to give more explicit explanation when lib32
is actually a mount point. However, the os.path.ismount() function
does not seem to detect bind mounts, as in the linked bug report.

Closes: https://bugs.gentoo.org/643632