dev-vcs/git: set PERL_PATH variable.
The rational for this change is that git on Gentoo/FreeBSD doesn't build
properly when this variable is unset. After chatting to floppym on the
gentoo-dev channel, here's what he found out in git-instaweb's Makefile:

Define PERL_PATH to the path of your Perl binary (usually /usr/bin/perl).

Package-Manager: Portage-2.3.13, Repoman-2.3.4
Closes: https://github.com/gentoo/gentoo/pull/6849