profiles: unset USE=modules by default.
The "modules" USE flag was originally enabled by default in the base
profile so that any ebuild inheriting linux-mod.eclass would have its
kernel modules built by default. However, the name of that USE flag is
now controlled by the MODULES_OPTIONAL_USE variable, set in the ebuild
itself. The best place for the USE flag default is therefore in IUSE
(in the ebuild), because only the ebuild knows the correct flag name.

There are at least two packages, media-gfx/graphicsmagick and
net-im/mcabber, that assign a different meaning to the "modules" flag,
further suggesting that a global default is too blunt an approach.

This commit drops the flag from base/make.defaults, and also from
default/linux/make.defaults where a redundant entry was present.
Thereafter, the override in features/hardened/make.defaults is
unnecessary, so it has been dropped as well.

Closes: https://bugs.gentoo.org/635720
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>