dev-lang/php: new revisions to fix shared memory and threads support.
Our PHP ebuilds used the "sharedmem" flag to conflate two ./configure
flags, namely --enable-shmop and --with-mm. The former is the "shared
memory" extension, and the latter enables the dev-libs/mm backend for
session storage. This commit separates the two:

  * The old USE=sharedmem enables the shared memory extension
  * The new USE=session-mm enables the dev-libs/mm session backend

The nascent session-mm flag depends on USE=session being set.

Additionally, we had a REQUIRED_USE constraint prohibiting USE=sharedmem
from being set at the same time as USE=threads. This turns out to be due
to an incompatibility between the --with-mm and --enable-maintainer-zts
flags, the latter of which is toggled by USE=threads. So in the new
revisions, we have a REQUIRED_USE constraint that blocks USE=threads
when USE=session-mm is set.

Thanks are due to Xiami who reported and fixed all of these issues in
bug 644922.

Reported-by: Xiami <i@f2light.com>
Closes: https://bugs.gentoo.org/644922
Package-Manager: Portage-2.3.13, Repoman-2.3.3