net-nds/rpcbind: use default values for "--bindir" and "--with-rpcuser"
Upstream commit 9afccfc moves the rpcbind executable from the bindir
to the sbindir, meaning that we don't have to override the bindir in
our ebuild if we want rpcbind installed to e.g. /usr/sbin. This commit
drops the "--bindir" flag from src_configure, so the default will be
used. As a side effect, this will place the rpcinfo executable in the
bindir where it belongs, allowing us to fix things like bug 457042.

Our ebuild was also passing "--with-rpcuser=root" explicitly, even
though that is the default value. This commit removes the redundancy.

Bug: https://bugs.gentoo.org/457042
Closes: https://bugs.gentoo.org/628766
Package-Manager: Portage-2.3.19, Repoman-2.3.6