Move current ruby23 status into arch profiles, bug 639476
Mask ruby_targets_ruby23 in the use.stable.mask of each arch and set
RUBY_TARGETS to ruby22 in the make.defaults for each arch.

This makes it possible to update the defaults in the main profile for
a stable ruby:2.3 slot and consequently to mark ruby:2.3 stable on
each arch in turn making sure that users of that arch will not end up
with conflicting settings (e.g. stable ruby:2.3 but
RUBY_TARGETS="ruby22").

Each arch is expected to remove these temporary entries as they mark
ruby:2.3 stable, as per the instructions in bug 639476.