dev-db/freetds: new version 1.00.341 from the "current" series.
This apparently-huge version jump, from 1.00.54 to 1.00.341, comes as
we switch from the upstream "stable" release series to the "current"
one. The stable series is just, well, too stable for us: there have
been a number of bug fixes (Gentoo bug 621430, for example) that never
made it into a stable release.

The "current" series should follow the upstream git repository more
closely; they're snapshots of one manner or another. It looks to me
like (and I have my fingers crossed that) the version numbers for the
stable and current series to not overlap.

Closes: https://bugs.gentoo.org/621430
Package-Manager: Portage-2.3.19, Repoman-2.3.6