net-misc/openssh: revbumps for net-libs/ldns dep change
Since sys-apps/portage-2.3.20 defaults to --dynamic-deps=n, this revbump
is needed in order to correct the dependencies of installed instances of
openssh.

It's not possible to revbump the 7.6_p1 ebuild to 7.6_p1-r1, since the
7.6_p1-r1 ebuild already exists with the only difference being that
the HPN_PATCH variable is commented out. As a compromise, make 7.6_p1-r2
include HPN_PATCH but also remove the default IUSE setting for hpn in
order to suppress bug 634594 by default.

Fixes: a71042d76cb4 ("net-misc/openssh: Fix net-libs/ldns dependency")
Bug: https://bugs.gentoo.org/645310
Bug: https://bugs.gentoo.org/634594
Package-Manager: Portage-2.3.20, Repoman-2.3.6
RepoMan-Options: --force