sys-libs/zlib: simplify build for mingw-w64
Currently the upstream configure script does not handle mingw-w64, with
a note to temporarily bypass it, indicating they are not opposed to a
change allowing it to work.

Patch configure and Makefile.in to accept mingw-w64 and produce
libraries with expected naming (pretty sure there are a lot of hardcoded
checks) and use configure for all CHOSTs.

Package-Manager: Portage-2.3.20, Repoman-2.3.6
Signed-off-by: Marty E. Plummer <hanetzer@startmail.com>