dev-scheme/kawa: Drop swt USE flag
This wasn't actually enabling SWT in the build as providing the jar
location actually disables it. There is no other way to provide the
jar location so enabling SWT simply breaks the build. The latest
version probably handles this better but no one cares about this
package.

Package-Manager: Portage-2.3.20, Repoman-2.3.6