x11-themes/lxqt-themes: new package
This repository comprises a number of graphic files and themes.
The LXQt logo was designed by @Caig and is licensed CC-BY-SA 3.0.
LXQt theme "Plasma" is based on Breeze 5.0.0 theme by the KDE Visual Team.
The Wallpaper is the contest winning design from Nuno Pinheiro
uno@oxygen-icons.org, licensed as LGPLv3.