sys-apps/portage: add gentoo-dev USE
The gentoo-dev USE flag will be use to enable stricter default
settings that are intended for use by ebuild developers.
This flag is not recommended for non-developers, since it
might enable settings that are not ready for general use.

This flag makes emerge default to --dynamic-deps=n.

https://bugs.gentoo.org/645550
https://bugs.gentoo.org/646458
Package-Manager: Portage-2.3.24, Repoman-2.3.6