dev-embedded/arduino: bump to 1.8.5
This is a major rewrite for a major bump. As we can see in the bug
referred below, a *lot* of work can be poured into this ebuild but we
can also see that having too large a scope makes us never bump. Thus, I
opted for a bump that didn't widen the scope of the v1.0.5 ebuild. We
can always improve upon this ebuild later.

This work is strongly inspired by maurerpe's fine work at
https://github.com/maurerpe/arduino-overlay/tree/master/dev-embedded/arduino-ide
but I scaled down the unbundling effort for simplicity's sake.

Closes: https://bugs.gentoo.org/525882
Closes: https://github.com/gentoo/gentoo/pull/7166
Package-Manager: Portage-2.3.19, Repoman-2.3.6