dev-util/sysprof: Switch maintenance to gnome@ as primary
sysprof is an optional dependency of gnome-builder and its previous
maintainers are all part of that. Given that gnome-builder is maintained
by gnome team, move sysprof to be maintained primarily by gnome@ as well.
Remove myself by name in the process, keep tetromino for now as he's not
me, and no need to remove at least an explicit CC unless he wants that.

Package-Manager: Portage-2.3.19, Repoman-2.3.6