profiles: p.mask old gnucash for now instead of outright removing them
I don't want to remove gnucash-2.6 without an explicit ack from
titanofold, but I do want to proceed with p.masking the vulnerable
webkit-gtk SLOTs. So put gnucash-2.6 in p.mask for now, until Aaron
removed it himself completely, hopefully soon. gnucash-docs is left
unmask for users that had it to still have it and upgrade to 3.0 docs
later on automatically (there don't appear to be any explicit gnucash
deps in gnucash-docs itself).

Bug: https://bugs.gentoo.org/629114