x11-libs/wxGTK: Add a 3.0.2.0-r4 straight to stable on arm that removes USE=webkit,gstreamer
Unfortunately arm is delayed with stabling 3.0.3, so I'm forced to introduce
a rebuild without USE=webkit of 3.0.2.0-r3 as r4 to stable arm users meanwhile,
to get rid of security vulnerable webkit-gtk slots.
As USE=gstreamer is only needed by packages that do not have any arm keywords,
also remove USE=gstreamer from this version, as it also uses security vulnerable
ancient versions of gstreamer (also fixed in 3.0.3 bump that's delayed for arm
stable), and we can just remove it now together with webkit, instead of hitting
the problem again when gstreamer:0.10 is all ready for security cleanup (and
having to do yet another revbump for this separately later) - if arm is still
delayed by then.

Bug: https://bugs.gentoo.org/643956
Bug: https://bugs.gentoo.org/629208
Package-Manager: Portage-2.3.19, Repoman-2.3.6