media-libs/mesa: depend on xdg-utils for USE=openmax to not break XDG specs
This is currently necessary due to the installation of 99mesaxdgomx with USE=openmax.
Ideally we wouldn't depend on xdg-utils as a whole with its perl deps, but
that's pending changed there. Unbreak things for the not so common case of
mesa USE=openmax for the time being.

Bug: https://bugs.gentoo.org/635040
Bug: https://bugs.gentoo.org/264647
Package-Manager: Portage-2.3.19, Repoman-2.3.6