x11: Move x11-proto/xorgproto to x11-base/xorg-proto
Once the individual x11-proto/ packages subsumed by xorg-proto are gone,
the x11-proto/ category will contain only xcb-proto, printproto (should
be tree cleaned) and xorgproto. That's not enough packages for a
category, so move xorgproto to x11-base/xorg-proto (with the hyphen to
match other packages in x11-base/).