dev-lang/smlnj: Bump to 110.82, fix install dirs, build on pax kernels.
Bump to 110.82, fixes 487064 , thanks to Keshav Kini and OGINO Masanori for
reporting.

Install to the smlnj subdirectory underneath the 32 bit lib directory if it
exists, or the 64 bit lib directory otherwise, and create symlinks in /usr/bin
to point to the executables.  Fixes 248162, thanks to flameeyes for reporting,
xmv and hkbst for helping.

Fix build on pax kernels, fixes 564132, thanks to Fedja Beader for reporting.

Closes: https://bugs.gentoo.org/487064
Closes: https://bugs.gentoo.org/248162
Closes: https://bugs.gentoo.org/564132
Package-Manager: Portage-2.3.24, Repoman-2.3.6