games-action/super-hexagon: Bump to 20160217, EAPI 6, no bundling
This previously bundled SDL 2, probably because it wasn't packaged
when this was first added. With that unbundled, there are far fewer
dependencies and this makes unbundling everything less likely to
break. SDL is particularly stable in this regard. Because of this and
in the interests of security, I have dropped the bundled-libs
flag. Anyone experiencing issues can obviously file a bug report.

Closes: https://bugs.gentoo.org/611282
Package-Manager: Portage-2.3.24, Repoman-2.3.6