media-video/obs-studio: Add support for scripting.
The configure automagic already enables support when the requisite
libraries are present.  This commit gives the user control on
whether or not to do so, and pulls in the required dependencies.

Currently support for both, Lua and Python will be enabled by the
automagic even if only one of them was enabled by the user, but only
if the libraries are found.

Package-Manager: Portage-2.3.24, Repoman-2.3.6