dev-db/myodbc: remove emul-linux-x86 references
Also cleaned up the patches and adjusted cmake-doc-path.patch to be -p1,
removed eutils as unneeded and switched epatch to eapply.

Versions myodbc-5.2.7-r1.ebuild and myodbc-5.2.7-r3.ebuild would not
build on my system, but failed in the exact same manner with the patches
not applied. Possibly they are too old to work on current gcc/glibc.

Package-Manager: Portage-2.3.24, Repoman-2.3.6