dev-python/google-apitools: drop cli/test/sample modules
Talking to upstream, the cli modules aren't generally useful, and gsutil
(the only user of this currently) doesn't need it.  By dropping that, we
can drop the deprecated google-apputils project which doesn't support
py3.  We also clean up the install of unused sample & test modules.  Now
we can make it support python-3.