sys-libs/glibc: Add support for building RISC-V headers
The RISC-V GCC port defines some extra preprocessor macros that glibc
uses to determine which RISC-V target is being compiled for.  Since
crossdev doesn't use a RISC-V compiler to build the headers, this blows
up.  These checks don't matter for building the headers because they're
the same on all RISC-V targets, so we just hack up the glibc build to
elide the check.