app-forensics/sleuthkit: bump to 4.6.0
This bump starts to bundle libewf since app-forensics/libewf is about
to get treecleaned (see #547418). The upstream only supports libewf
version 20130128[0], which is not available in the tree. Because they
haven't clarified the supported libewf versions until recently, we
have been depending on any version and it's been causing build
failures (see #607968). Although there are compatibility patches to
support later versions of libewf in tsk, they are not supported by
upstream. There's little to no expactation of tsk updating its code to
use the latest libewf since they've forked the version 20130128[1].

In terms of stability, 20130128 was marked stable in Gentoo at some
point[2]. There are no known security vulnerabilities. If in the
future the upstream fork diverges, we can add it to the tree as a new
package and establish a proper dependency relationship.

Note though that the ewf USE flag is masked by treecleanears[3], so
this change currently has no visible impact on users.

[0] https://github.com/sleuthkit/sleuthkit/blob/sleuthkit-4.6.0/INSTALL.txt#L44
[1] https://github.com/sleuthkit/libewf_64bit
[2] https://sources.gentoo.org/cgi-bin/viewvc.cgi/gentoo-x86/app-forensics/libewf/libewf-20130128.ebuild?revision=1.6&view=markup
[3] https://gitweb.gentoo.org/repo/gentoo.git/tree/profiles/base/package.use.mask?id=f103062521b15cddc67a822a7a25640d3fbab76a#n65

Bug: https://bugs.gentoo.org/547418
Bug: https://bugs.gentoo.org/607968
Package-Manager: Portage-2.3.27, Repoman-2.3.9