mail-client/neomutt: Bump live ebuild version
On unstable systems (or once a regular version of neomutt is made
stable) the live ebuild version of '9999' will not be the latest version
and a user will not be able to easily install the live ebuild. This
change renames it to `99999999`.

Thanks to contact@ryper.org for pointing this out!

Closes: https://bugs.gentoo.org/650434
Package-Manager: Portage-2.3.24, Repoman-2.3.6