app-arch/bzip2: EAPI update and patch cleanup.
Think of this as a precursor to switching bzip2 to an autotools build.
EAPI has been bumped from 5 to 6, and the patches have been reformatted
to apply with eapply in in addition to epatch (which apparently tries
every -pN option with patch until it hits one that works), so the new
revbump can be tested and stabilized while the old version is still
available without change.

Also changed dodoc/dohtml to DOCS=()/HTML_DOCS=() einstalldocs.

Package-Manager: Portage-2.3.28, Repoman-2.3.9
Closes: https://github.com/gentoo/gentoo/pull/7848