sys-devel/binutils: fix ldscrips ${CTARGET} search path, bug #147155
This change fixes long-standing search path issue in Gentoo's binutils:

Before the change search path was the following:
    /usr/${CTARGET}/lib/ldscripts
Note: it points to $SYSROOT, not to native cross-tools.

After the change search path is the following:
    /usr/${CHOST}/${CTARGET}/lib/ldscripts

Added two notes to the ebuild on how things are supposed to work:
- Note [slotting support]
- Note [tooldir hack for ldscripts]

Applied change to 2.30-r1 and live ebuilds.

Reported-by: Heiko Rosemann
Closes: https://bugs.gentoo.org/147155
Package-Manager: Portage-2.3.28, Repoman-2.3.9