net-libs/libsearpc: Fix SRC_URI
Upstream has an insane usage of git tags, so the tag is called
v3.1-latest, but it is actually the library version 3.0.8.
On the next version bump, we at least need to change the SRC_URI to use
the snapshot tarball, but right now this would lead to mirrors serving
the wrong tarball, which is probably not the best thing to do.

Closes: https://bugs.gentoo.org/652760