net-misc/iputils: fix ipv6 dependency logic and drop REQUIRED_USE
This revision bump adjusts the logic for OpenSSL  and LibreSSL ensuring
that the libressl flag is not dependent on the openssl flag being set.

Also, this revision bump removes the requirement for REQUIRED_USE and
defaults to dev-libs/openssl if no user flags are set. OpenSSL was
chosen as it is the current default provider for a stage3.  Should the
libressl project decide to build a stage3 this can also be easily
adjusted in profiles without ebuild modifications.

Bug: https://bugs.gentoo.org/643304
Package-Manager: Portage-2.3.28, Repoman-2.3.9
Closes: https://github.com/gentoo/gentoo/pull/7937