dev-lang/erlang: Version bump to 20.3.2
This change includes a major update to the ebuild. I, with guidance from gokturk, have fixed many mistakes and
cleaned up the ebuild. The build system is difficult to work with, and requires some special autoconf work to get the
correct configure scripts generated. In addition, please note that the erlang.org and github releases for the same version
are different, the biggest factor being that the erlang.org release includes precompiled binaries, while the github
release does not.

Please also note that I have dropped the "wxwidgets" USE flag, as upstream unconditionally links against it

Signed-off-by: Nick Sarnie <commendsarnex@gmail.com>

Package-Manager: Portage-2.3.28, Repoman-2.3.9