net-libs/gtk-vnc-0.7.2: remove pygtk2 based python binding
These are now deprecated upstream to the point of outputting a fat warning
at end of configure and NEWS file says they will be deleted for next
version.
The only in-tree consumer of the python USE flag was a mistake and
has been fixed appropriately. As such, just remove the support fully with
the bump, which simplifies the ebuild quite a bit. Also next commit will
remove gtk2 support, which these python bindings are based on.

Introspection (supported by gtk-vnc[introspection]) is the modern way
to use gtk-vnc from python, and this works with python2 as well.

Package-Manager: Portage-2.3.28, Repoman-2.3.9