dev-util/spirv-tools: Restrict tests
The tests go through all symbols in the generated libraries/executables, and check the symbol names against a hardcoded regular expression.
If an unmatching symbol is found, the test fails. This is easily hit if a user sets a compiler flag that causes symbols to change, such as PIC

Fixes: https://bugs.gentoo.org/651626

Package-Manager: Portage-2.3.31, Repoman-2.3.9