dev-games/ogre: New snapshot ebuild for the 1.11 series.
This is an ebuild for the last commit in the default branch from 2018-04-23.

I have decided to not provide a true live ebuild, as there are many commits
from which Ogre3D can not be built.

Although a build failure of a live package would only need that the user can
try later, and maybe succeed, this render needs a considerable amount of
compile time.

I will update the ebuild to a later commit (and revision), whenever I see
significant commits being made, and can ensure, that the resulting source tree
can be built.

Bug: https://bugs.gentoo.org/647266
Tested-by: Sven Eden yamakuzure@gmx.net

Package-Manager: Portage-2.3.28, Repoman-2.3.9