dev-db/mysql-init-scripts: Respect my.cnf user setting
**Blocker to merging:** How do we set the user in the s6 init?

* Fix running mysql as a different user
* Fix multiple Galera SST failures when /proc is mounted with hidepid=2.

On startup wsrep_sst_rsync was failing to run because it checks to
see if the mysqld process is running. When the service is started
without --user, the mysqld process calls setuid() to perform the
user change itself. The setuid call will flip the dumpable bit for
the process, disabling the ability for the mysql user to ptrace it.

The hidepid=2 /proc mount option checks not only if the process is
owned by the user, but also if the user can ptrace the process. If a
process can not be ptraced, it does not appear in the output of ps.

As a result of this, wsrep_sst_rsync would start the rsync daemon
and immediately kill it because it thought mysqld wasn't running,
since it didn't appear in ps.

Starting mysqld as the correct user allows the process to appear.

Closes: https://bugs.gentoo.org/593046
Package-Manager: Portage-2.3.24, Repoman-2.3.6