python-utils-r1.eclass: Initial EAPI 7 support
Add initial EAPI 7 support.  Handle trailing slash in D/ED correctly.
Support new has_version switches.

This does not provide proper cross support.  In particular, the PYTHON
variable (full path to the interpreter) is used both in contexts
of BDEPEND and DEPEND, so we need to clean the uses first.