dev-cpp/glibmm: Fix build with GCC 8
Patch taken from upstream and adapted to actually touch the pregenerated
file shipped with the release tarbal as well. That code appears to have
been always wrong, with no chance to build. This probably means that GCC
8 is suddenly being detected in some other way, making the code use that
include file which was previously apparently kept unused. But anyway,
upstream killed that include in a later release, so let's just let this
build and watch the eventual breakage.

Upstream says:

> Fixed in the glibmm-2-54 branch. No fix is necessary in the master
> branch. The threads.hg and threads.h files don't exist there.

Closes: https://bugs.gentoo.org/654776
Bug: https://bugzilla.gnome.org/show_bug.cgi?id=791711
Closes: https://github.com/gentoo/gentoo/pull/8324