www-client/chromium: widevine rehabilitation
>chromium-66 ebuilds dropped the widevine useflag because upstream
decided (again! but only "for now") to add code to induce build
failure when linux builds request widevine.

However, it seems that if we just add linux to the whitelist,
everything works as before (but buckle up, there may be
turbulent skies ahead).

nb: ninja no longer copies the widevine .so to out/ (so we needn't
remove it anymore in src_install).

Signed-off-by: Gregory M. Turner <gmt@be-evil.net>