scons-utils.eclass: Provide proper Python API for EAPI 7
Provide a proper multi-impl Python support for scons-utils in EAPI 7,
to account for new versions of dev-util/scons (3.0.1-r100+, to be
committed) that support Python 3 and break SConstruct files using
Python 2 constructs.

Combining scons-utils with python-any-r1 and python-single-r1 is added
retroactively for older EAPIs as well, with fallback to Python 2.7.
The new (hard-to-use) API for python-r1 is specific to EAPI 7 since it
requires adding explicit BDEPEND.

The new use of the eclass is described on the wiki page, along with
series of examples covering different use cases:
https://wiki.gentoo.org/wiki/Project:Python/scons-utils_integration