Fix up init script and make this slightly more secure
This adds a better init script that actually works now and uses config
from /etc/conf.d during service start. Instead of placing the ssh sock
in a directory that is world readable we now make a user and group that
people who can access the ssh_sock should be added to.