sys-apps/flashrom: Add Atmel AT25DF021A support via patch
The Atmel AT25DF021A is used for the FPGA firmware in Raptor Talos II POWER9 workstations.

The patch is controlled by a USE flag since it is non-standard and adds a SRC_URI.

The ebuild revision is not bumped, because it is only ~arch, has a USE flag change, and is identical to the prior package without that USE flag set.