dev-util/debugedit: reversion current stable #653906
We want to switch release streams from the RPMv5 project to the RedHat
RPM project on GitHub as the RPMv5 project is dead now.  Unfortunately,
the RH versions are in the 4.x range, and we're already at 5.x, so we
need to move this to something smaller for unstable to upgrade to.

Add a "0." prefix for existing users (stable & unstable) to "downgrade"
to so the new 4.x series will be an upgrade for everyone.  It means a
useless rebuild, but this package is tiny, so shouldn't be a big deal.

Bug: https://bugs.gentoo.org/653906