toolchain-glibc.eclass: remove git logic
The git logic is handled in the live ebuild directly, and the only use
of this eclass is in non-live ebuilds. In fact, the two newest non-live
ebuilds do not make use of this eclass at all, and as far as I can see
nowhere in the history of this file does it inherit git-2 at all.