www-apps/grafana: fix pid file permission.
Due to a fact that the service is running as a user named grafana, the process
did not have permissions to create a file in /run. It could be mitigated by
changing permissions of /run or running service as privileged user or avoiding
systemd.  This patch moves pid file to subdirectory and adds RuntimeDirectory
entry to the definition of the service.

Closes: https://bugs.gentoo.org/628328

Package-Manager: Portage-2.3.19, Repoman-2.3.6
Closes: https://github.com/gentoo/gentoo/pull/8765