dev-libs/glib-2.54: handle py glib-{mkenums,genmarshal} more properly
glib-2.54 ported glib-mkenums and glib-genmarshal to python. Handle them
more properly via python_replicate_script and PYTHON_DEPS, so we actually
ensure that the deps are in place for the tools to work.
This is meant to be temporary and not a stable candidate until at least
these python tools are moved out into a separate package, like
gdbus-codegen is. But we are delayed in providing a newer glib for
~arch long enough now, so go with the simpler way to start with, as
separate package would involve consumer transition as well, including
addition of the new package into various packages DEPEND, which is hard
to know when it's needed (probably requiring some sort of QA check).

Bug: https://bugs.gentoo.org/651830
Package-Manager: Portage-2.3.40, Repoman-2.3.9