dev-libs/libsavitar: use pugixml from system resolving file collision
Upstream put pugixml inside the distribution and it is built and installed
alongside with the library. This results in file collisions when pugixml is
already installed on the system.

Existing upstream pull request: https://github.com/Ultimaker/libSavitar/pull/8

This commit applies that pull request as a patch. I modified the patch slightly
to not include the actual deletion of the library and only apply the needed
changes to the rest of the library. I did this because otherwise the filesize
was bigger than the portage limit.

Closes: https://bugs.gentoo.org/658112
Package-Manager: Portage[mgorny]-2.3.36.1