games-action/trine2: Version bump to 2.01.425, Complete Story edition
This edition includes the Goblin Menace expansion. Sorry if you only
have the original but this was about to be last-rited and I don't have
the original to test against. You cannot buy it any more either.

It's a little buggy regarding displays and sound but once it's
running, you should be okay. Blame the short-lived SDL 1.3. If it
bothers you, poke Frozenbyte about rebuilding it against SDL 2. They
already gave Trine Enchanted Edition the same treatment so it
shouldn't be that hard.

Closes: https://bugs.gentoo.org/485058
Package-Manager: Portage-2.3.40, Repoman-2.3.9