net-analyzer/nagios-core: finally drop nagios-3.x from the tree.
A bunch of security bugs (and my own personal happiness) were blocked
on this because nagios-3.x was sometimes required by an old version of
net-analyzer/pnp4nagios. Now that a version of pnp4nagios has been
stabilized with more reasonable dependencies, we can finally get rid
of nagios-3.x.

This leaves a bunch of unused files in FILESDIR that I'll clean up in
another commit.

Bug: https://bugs.gentoo.org/636234
Bug: https://bugs.gentoo.org/628086
Bug: https://bugs.gentoo.org/629380
Package-Manager: Portage-2.3.40, Repoman-2.3.9