profiles/prefix/make.defaults: drop FEATURES=collision-protect
After many years, we finally came to the conclusion we don't need this
any more, and can do with protect-owned (the default) very well.

Most prominent reason for this is that Gentoo for Mac OS X has been
superseeded by Gentoo Prefix for macOS, which no longer tries to install
binaries side by side with the host provided utilities.  E.g. it is no
longer the case that files encountered as a conflict can be something
that is provided by the host system (and hence should not be nuked).

In Gentoo Prefix we assume that $EPREFIX is solely managed by the PM, so
anything it finds there which it doesn't know about can be safely
overwritten.  Users that actually use EPREFIX=/ on their systems can set
FEATURES=collision-protect in their local make.conf.

Bug: https://bugs.gentoo.org/655414