net-misc/openssh: update set of locale vars #367017
Rather than glob all LC_* env vars, hardcode the list based on what
POSIX defines and the GNU extensions.  In practice, this means we
additionally send LANGUAGE, and unregistered or unrelated env vars
that happen to start with "LC_" are no longer sent.  If someone has
a locale related variable that they want to include, they can file
a new request for it.

We continue to omit POSIX's NLSPATH as that is much more likely to
be system specific and have no useful meaning on the remote side.

Bug: https://bugs.gentoo.org/367017