dev-python/gst-python: bump to 1.14.1, run tests, don't install plugin
The plugin must be single-python, otherwise multiple plugin files (due to
different python plugin naming suffixes) get installed, gst-inspect-1.0 and
co will load both into memory, end up with multiple libpython.so's loaded in,
causing ABI issues and crashes from clashing symbols. python-r1 and
python-single-r1 can't co-exist, so stop building and installing the plugin
in this package, as this package is mainly about the GI overrides for things
like correct Gst.Fraction, and is consumed as such. The plugin is of limited
use, as that's about supporting writing and using gstreamer plugins in
python, but GIL gets in the way of writing good plugins, so it's not used
much in practice. Debian didn't even build the plugin before 1.14 either.
If it's ever requested/needed, we'll package it as
media-plugins/gst-plugins-python that build and installs just the plugin/
directory.

Fix tests to work in the multibuild setup and properly xdg_environment_reset,
which was missing and made apparent by tests printing an ERROR about it
(though passing).

Remove redundant IUSE=test

Use ltprune.eclass instead of eutils.eclass

Package-Manager: Portage-2.3.40, Repoman-2.3.9