perl-core/Archive-Tar: Add 2.300.0 for CVE-2018-12015 bug #657968
This resurrects perl-core/Archive-Tar that was previously removed
from tree, so that it can be installed independently on top of Perl
5.24 and perl 5.26, without forcing users to upgrade to perl 5.28
just to get this fix.

virtual/perl-Archive-Tar will be also mapped for this purpose.

Keywords taken from dev-lang/perl-5.26.2 on the basis that due to
Archive-Tar being included there, that this codebase is already
keyworded under a different name.

Patches applied to fence out Text-Diff dependent mechanics and a bunch
of other weirdness that seem to be targeted for non-vendor layouts
( eg: it assumes we're installing bin/* into /usr/local/bin )

I'm not 100% on the mechanics I've chosen, but its OK for now.

Package-Manager: Portage-2.3.40, Repoman-2.3.9