dev-perl/Archive-Zip: Add fix for CVE-2018-10860 bug #660466
This includes upstreams fixes and tests including binary files
presented as textual diffs, which appears to work.

Upstream testing indicates there are potential issues with tests on
some systems, and so it may not be suitable as-is for stabilization

See Github PR mentioned below for additional details.

Bug: https://bugs.gentoo.org/660466
Bug: https://bugzilla.redhat.com/show_bug.cgi?id=1591449
Bug: https://github.com/redhotpenguin/perl-Archive-Zip/pull/33
Package-Manager: Portage-2.3.40, Repoman-2.3.9