dev-python/pillow: fix broken build process
In 4.3.0 and 5.2.0, we had a broken way of passing flags to Pillow's
build script. Flags would be passed fine in the compile phase, but would
not be passed at all at the install phase, whichwas the equivalent of
not sending flags at all.

I'm committing straight to stable to 4.3.0-r2 because the build process
is broken.

Closes: https://bugs.gentoo.org/661308
Package-Manager: Portage-2.3.43, Repoman-2.3.10