net-libs/webkit-gtk-2.20.3: disable experimental accelerated 2D canvas
ACCELERATED_2D_CANVAS is considered experimental by upstream. I'm not sure
what the justification was for enabling it with USE="webgl opengl" for us.
The only reason this code isn't removed fully from upstream is because it
_MIGHT_ be useful for some embedded use cases, yet it doesn't work with
GLES2 yet, only desktop opengl.
Additionally, for us this is causing additional cairo[opengl] dependency
that can't be avoided for just non-experimental WebGL support via desktop
GL as it was setup before in ebuild. No-one but Gentoo is building with
this experimental option (not even upstream developers), and any issues
about it won't really be cared about.
Disable it and drop the cairo[opengl] dependency.

https://lists.webkit.org/pipermail/webkit-gtk/2017-March/002969.html
https://lists.webkit.org/pipermail/webkit-gtk/2017-April/003004.html

If this is ever added back for any reason, make sure cairo[opengl] isn't
depended on when not necessary (when accelerated 2d canvas is disabled,
e.g. previously USE="opengl gles2" pulled in cairo[opengl] still, albeit
that combination was not allowed previously, but a future webkit-gtk
upstream version may support them together) and review if cairo-glesv2
is supported (and so it'd be supported with USE=gles2 too, unlike now).

Bug: https://bugs.gentoo.org/581248
Closes: https://bugs.gentoo.org/643726
Package-Manager: Portage-2.3.43, Repoman-2.3.10